+++ apps/engine/core/routes.py
@@ -1,8 +1,8 @@
 from flask import Blueprint, jsonify
 
-from app.engine.core.services.thingspeaks_service import ThingSpeaksService
-from app.engine.core.services.engine_service import EngineService
-from app.engine.core.enums.http_status import HttpStatus
+from apps.engine.core.services.thingspeaks_service import ThingSpeaksService
+from apps.engine.core.services.engine_service import EngineService
+from apps.engine.core.enums.http_status import HttpStatus
 from flask import jsonify, request
 
 """
